feat: add docs for npm and templates (#86)
* feat: add docs for npm and templates

* Specify code block language in installation docs

* Revamp getting started guide and starter templates

* Update supported-platforms.md

* Update starter templates and improve headings

* Update getting-started.md

* Update getting-started.md

* Update getting-started.md

* Capitalize NPM

* Update usage examples in getting-started guide

this was based on a short chat between me and daniel deciding that each example should be runnable as is. so simplified a bit

* Add Svelte starter template to getting started guide

* Removed `recommended` approach

I'm removing this for now because it can cause issues with ai and vibe-coding tools that rely on one-shotting an app using the cdn approach. we can bring it back once we explain clearly when to use each.

* Add Astro starter template and update labels

* Update supported-platforms.md

* Update getting-started.md

---------

Co-authored-by: jelveh <nj@puter.com>
Co-authored-by: Reynaldi Chernando <reynaldichernando@gmail.com>